Extract api code from `PlotView`
* Move `useQuery` hook into a dedicated hook `useReadings`
* Clean up code and simplify
* Update README
